Fix logprob sample saving for BFloat16 tensors
Numpy doesn't support BFloat16, so convert to float32 before calling
.cpu().numpy(). This fixes the TypeError when saving logprob samples
without fp32_lm_head enabled.

+++ open_instruct/grpo_fast.py
@@ -994,8 +994,8 @@ class PolicyTrainerRayProcess(RayProcess):
                         # Save raw logprob samples for alignment analysis (ScaleRL Figure 3)
                         if self.args.save_logprob_samples:
                             self._save_logprob_samples(
-                                vllm_logprobs_BT[valid_mask_BT].cpu().numpy(),
-                                local_logprobs_BT[valid_mask_BT].cpu().numpy(),
+                                vllm_logprobs_BT[valid_mask_BT].float().cpu().numpy(),
+                                local_logprobs_BT[valid_mask_BT].float().cpu().numpy(),
                             )
 
                     new_logprobs_BT = local_logprobs_BT
